requirements.txt
streamlit-authenticator
bs4
lxml
python-pptx
st-star-rating